Quota PUT endpoint: Add l7policy parameter
Since Limes is now talking directly to Octavia, it is (always has
been) trying to set quotas for l7policies, among others.
Octavia stein does not allow this field though, so it errors out.

We don't want to add an exception to Limes, that would be removed
again anyway as soon as Octavia is updated (to a version supporting
l7policy quotas), so instead we just allow the l7policy field, but
ignore it.